cmd/compile: change Type.Alignment to return uint8
This makes it match Type.Align,
which will ease refactoring to call dowidth automatically.

+++ src/cmd/compile/internal/gc/type.go
@@ -916,9 +916,9 @@ func (t *Type) Size() int64 {
 	return t.Width
 }
 
-func (t *Type) Alignment() int64 {
+func (t *Type) Alignment() uint8 {
 	dowidth(t)
-	return int64(t.Align)
+	return t.Align
 }
 
 func (t *Type) SimpleString() string {
